@@ -3,55 +3,95 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
     <groupId>com.zrlog</groupId>
-    <artifactId>plugin-email</artifactId>
-    <version>1.2.0</version>
+    <artifactId>plugin-mail</artifactId>
+    <version>3.0.23</version>
     <packaging>jar</packaging>
+
     <properties>
-        <outputDirectory>/home/xiaochun/zrlog-plugin/</outputDirectory>
+        <outputDirectory>/Users/xiaochun/plugins/</outputDirectory>
+        <imageName>mail</imageName>
+        <native.maven.plugin.version>0.10.2</native.maven.plugin.version>
+        <native.image.mainClass>com.zrlog.plugin.mail.Application</native.image.mainClass>
+        <native.image.agentMainClass>com.zrlog.plugin.mail.GraalvmAgentApplication</native.image.agentMainClass>
     </properties>
+
 
     <dependencies>
         <dependency>
-            <groupId>com.zrlog</groupId>
-            <artifactId>plugin-common</artifactId>
-            <version>1.3.0</version>
+            <groupId>com.hibegin</groupId>
+            <artifactId>zrlog-plugin-common</artifactId>
+            <version>3.0.23</version>
         </dependency>
         <dependency>
-            <groupId>javax.mail</groupId>
-            <artifactId>mail</artifactId>
-            <version>1.4.7</version>
+            <groupId>com.sun.mail</groupId>
+            <artifactId>jakarta.mail</artifactId>
+            <version>2.0.1</version>
         </dependency>
     </dependencies>
 
     <build>
-        <finalName>mail</finalName>
+        <finalName>oss</finalName>
         <resources>
             <resource>
                 <directory>src/main/resources</directory>
                 <includes>
-                    <include>**/**</include>
+                    <include>**/*.properties</include>
+                    <include>*.base64</include>
+                    <include>**/templates/**</include>
                 </includes>
             </resource>
         </resources>
         <plugins>
             <plugin>
+                <groupId>org.codehaus.mojo</groupId>
+                <artifactId>exec-maven-plugin</artifactId>
+                <version>3.1.0</version>
+                <executions>
+                    <execution>
+                        <id>java-agent</id>
+                        <goals>
+                            <goal>exec</goal>
+                        </goals>
+                        <configuration>
+                            <executable>java</executable>
+                            <workingDirectory>${project.build.directory}</workingDirectory>
+                            <arguments>
+                                <argument>--enable-preview</argument>
+                                <argument>-classpath</argument>
+                                <classpath/>
+                                <argument>${native.image.agentMainClass}</argument>
+                            </arguments>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>native</id>
+                        <goals>
+                            <goal>exec</goal>
+                        </goals>
+                        <configuration>
+                            <executable>${project.build.directory}/${imageName}</executable>
+                            <workingDirectory>${project.build.directory}</workingDirectory>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
                 <artifactId>maven-compiler-plugin</artifactId>
                 <version>3.1</version>
                 <configuration>
                     <encoding>UTF-8</encoding>
-                    <source>1.7</source>
-                    <target>1.7</target>
+                    <source>1.8</source>
+                    <target>1.8</target>
                 </configuration>
             </plugin>
             <plugin>
                 <artifactId>maven-assembly-plugin</artifactId>
-                <version>2.5.5</version>
+                <version>3.1.0</version>
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>com.zrlog.plugin.mail.Start</mainClass>
+                            <mainClass>com.zrlog.plugin.mail.Application</mainClass>
                         </manifest>
                     </archive>
                     <finalName>mail</finalName>
