@@ -1,12 +1,12 @@
 package com.zrlog.plugin.mail.util;
 
-import javax.activation.DataHandler;
-import javax.activation.FileDataSource;
-import javax.mail.*;
-import javax.mail.internet.InternetAddress;
-import javax.mail.internet.MimeBodyPart;
-import javax.mail.internet.MimeMessage;
-import javax.mail.internet.MimeMultipart;
+import jakarta.activation.DataHandler;
+import jakarta.activation.FileDataSource;
+import jakarta.mail.*;
+import jakarta.mail.internet.InternetAddress;
+import jakarta.mail.internet.MimeBodyPart;
+import jakarta.mail.internet.MimeMessage;
+import jakarta.mail.internet.MimeMultipart;
 import java.io.File;
 import java.util.*;
 
@@ -16,11 +16,10 @@ public class MailUtil {
     private MailUtil() {
     }
 
-    public static boolean sendMail(String to, String title, String content, Map<String, Object> sMTPMap, List<File> files)
-            throws Exception {
+    public static boolean sendMail(String to, String title, String content, Map<String, Object> smtpMap, List<File> files) throws Exception {
         List<String> tos = new ArrayList<>();
         tos.add(to);
-        return sendMail(tos, title, content, sMTPMap, files);
+        return sendMail(tos, title, content, smtpMap, files);
     }
 
     public static boolean sendMail(List<String> to, String title, String content, Map<String, Object> sMTPMap, List<File> files) throws Exception {
@@ -33,8 +32,7 @@ public class MailUtil {
         };
         final Session session = Session.getDefaultInstance(prop, sMTPAuth);
         final Message message = new MimeMessage(session);
-        Address address = new InternetAddress(
-                prop.getProperty("mail.smtp.from"), displayName);
+        Address address = new InternetAddress(prop.getProperty("mail.smtp.from"), displayName);
         InternetAddress[] addresses = new InternetAddress[to.size()];
         for (int i = 0; i < to.size(); i++) {
             addresses[i] = new InternetAddress(to.get(i));
@@ -61,37 +59,34 @@ public class MailUtil {
         message.setContent(mp);
         message.setSentDate(new Date());
         message.saveChanges();
-        new Thread() {
-            public void run() {
-                try {
-                    Transport tran = session.getTransport("smtp");
-                    tran.connect(prop.getProperty("mail.smtp.host"),
-                            prop.getProperty("mail.smtp.username"),
-                            prop.getProperty("mail.smtp.password"));
-                    tran.sendMessage(message, message.getAllRecipients());
-                    tran.close();
-                } catch (MessagingException e) {
-                    e.printStackTrace();
-                }
+        Transport tran = null;
+        try {
+            tran = session.getTransport("smtp");
+            tran.connect(prop.getProperty("mail.smtp.host"), prop.getProperty("mail.smtp.username"), prop.getProperty("mail.smtp.password"));
+            tran.sendMessage(message, message.getAllRecipients());
+            tran.close();
+        } finally {
+            if (Objects.nonNull(tran)) {
+                tran.close();
             }
-        }.start();
+        }
         return true;
     }
 
-    public static boolean sendMail(List<String> to, String title, String content, Map<String, Object> sMTPMap)
-            throws Exception {
-        return sendMail(to, title, content, sMTPMap, new ArrayList<File>());
-    }
 
-    private static Properties getMailSettings(Map<String, Object> sMTPMap) {
+    private static Properties getMailSettings(Map<String, Object> smtpMap) {
         Properties properties = new Properties();
-        properties.setProperty("mail.smtp.displayName", "ZrLog系统邮件");
+        String displayName = (String) smtpMap.get("displayName");
+        if (Objects.isNull(displayName) || displayName.isEmpty()) {
+            displayName = "ZrLog 系统邮件";
+        }
+        properties.setProperty("mail.smtp.displayName", displayName);
         properties.setProperty("mail.smtp.auth", "true");
-        properties.setProperty("mail.smtp.username", sMTPMap.get("from").toString());
-        properties.setProperty("mail.smtp.port", sMTPMap.get("port").toString());
-        properties.setProperty("mail.smtp.password", sMTPMap.get("password").toString());
-        properties.setProperty("mail.smtp.host", sMTPMap.get("smtpServer").toString());
-        properties.setProperty("mail.smtp.from", sMTPMap.get("from").toString());
+        properties.setProperty("mail.smtp.username", smtpMap.get("from").toString());
+        properties.setProperty("mail.smtp.port", smtpMap.get("port").toString());
+        properties.setProperty("mail.smtp.password", smtpMap.get("password").toString());
+        properties.setProperty("mail.smtp.host", smtpMap.get("smtpServer").toString());
+        properties.setProperty("mail.smtp.from", smtpMap.get("from").toString());
         return properties;
     }
 }
