@@ -1,21 +1,21 @@
 package com.zrlog.plugin.mail.controller;
 
-import com.zrlog.plugin.IMsgPacketCallBack;
+import com.google.gson.Gson;
 import com.zrlog.plugin.IOSession;
 import com.zrlog.plugin.common.IdUtil;
 import com.zrlog.plugin.common.LoggerUtil;
+import com.zrlog.plugin.common.model.PublicInfo;
 import com.zrlog.plugin.data.codec.ContentType;
 import com.zrlog.plugin.data.codec.HttpRequestInfo;
 import com.zrlog.plugin.data.codec.MsgPacket;
 import com.zrlog.plugin.data.codec.MsgPacketStatus;
 import com.zrlog.plugin.mail.util.MailUtil;
 import com.zrlog.plugin.type.ActionType;
-import com.google.gson.Gson;
 
-import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Objects;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -24,11 +24,11 @@ import java.util.logging.Logger;
  */
 public class MailController {
 
-    private static Logger LOGGER = LoggerUtil.getLogger(MailController.class);
+    private static final Logger LOGGER = LoggerUtil.getLogger(MailController.class);
 
-    private IOSession session;
-    private MsgPacket requestPacket;
-    private HttpRequestInfo requestInfo;
+    private final IOSession session;
+    private final MsgPacket requestPacket;
+    private final HttpRequestInfo requestInfo;
 
     public MailController(IOSession session, MsgPacket requestPacket, HttpRequestInfo requestInfo) {
         this.session = session;
@@ -36,52 +36,42 @@ public class MailController {
         this.requestInfo = requestInfo;
     }
 
-    public void info() {
-        Map<String, Object> keyMap = new HashMap<>();
-        keyMap.put("key", "to,from,smtpServer,password,port");
-        session.sendJsonMsg(keyMap, ActionType.GET_WEBSITE.name(), IdUtil.getInt(), MsgPacketStatus.SEND_REQUEST, new IMsgPacketCallBack() {
-            @Override
-            public void handler(MsgPacket msgPacket) {
-                Map map = new Gson().fromJson(msgPacket.getDataStr(), Map.class);
-                session.sendMsg(new MsgPacket(map, ContentType.JSON, MsgPacketStatus.RESPONSE_SUCCESS, requestPacket.getMsgId(), requestPacket.getMethodStr()));
-            }
-        });
-    }
-
     public void update() {
-        session.sendMsg(new MsgPacket(requestInfo.simpleParam(), ContentType.JSON, MsgPacketStatus.SEND_REQUEST, IdUtil.getInt(), ActionType.SET_WEBSITE.name()), new IMsgPacketCallBack() {
-            @Override
-            public void handler(MsgPacket msgPacket) {
-                Map<String, Object> map = new HashMap<>();
-                map.put("success", true);
-                session.sendMsg(new MsgPacket(map, ContentType.JSON, MsgPacketStatus.RESPONSE_SUCCESS, requestPacket.getMsgId(), requestPacket.getMethodStr()));
-            }
+        session.sendMsg(new MsgPacket(requestInfo.simpleParam(), ContentType.JSON, MsgPacketStatus.SEND_REQUEST, IdUtil.getInt(), ActionType.SET_WEBSITE.name()), msgPacket -> {
+            Map<String, Object> map = new HashMap<>();
+            map.put("success", true);
+            session.sendMsg(new MsgPacket(map, ContentType.JSON, MsgPacketStatus.RESPONSE_SUCCESS, requestPacket.getMsgId(), requestPacket.getMethodStr()));
         });
     }
 
     public void index() {
-        session.responseHtml("/templates/index.html", new HashMap(), requestPacket.getMethodStr(), requestPacket.getMsgId());
-
+        Map<String, Object> keyMap = new HashMap<>();
+        keyMap.put("key", "to,from,smtpServer,password,port");
+        session.sendJsonMsg(keyMap, ActionType.GET_WEBSITE.name(), IdUtil.getInt(), MsgPacketStatus.SEND_REQUEST, msgPacket -> {
+            Map map = new Gson().fromJson(msgPacket.getDataStr(), Map.class);
+            Map<String, Object> data = new HashMap<>();
+            data.put("theme", Objects.equals(requestInfo.getHeader().get("Dark-Mode"), "true") ? "dark" : "light");
+            data.put("data", new Gson().toJson(map));
+            session.responseHtml("/templates/index.html", data, requestPacket.getMethodStr(), requestPacket.getMsgId());
+        });
     }
 
     public void testEmailService() {
         Map<String, Object> keyMap = new HashMap<>();
         keyMap.put("key", "to,from,smtpServer,password,port");
-        session.sendJsonMsg(keyMap, ActionType.GET_WEBSITE.name(), IdUtil.getInt(), MsgPacketStatus.SEND_REQUEST, new IMsgPacketCallBack() {
-            @Override
-            public void handler(MsgPacket msgPacket) {
-                Map map = new Gson().fromJson(msgPacket.getDataStr(), Map.class);
-                Map<String, Object> response = new HashMap<>();
-                try {
-                    MailUtil.sendMail(map.get("to").toString(), "这是一封测试邮件", "<h3>这是一封测试邮件</h3>\n", map, new ArrayList<File>());
-                    response.put("status", 200);
-                } catch (Exception e) {
-                    LOGGER.log(Level.SEVERE,"send email error ", e);
-                    response.put("status", 500);
-                }
-                session.sendMsg(ContentType.JSON, response, requestPacket.getMethodStr(), requestPacket.getMsgId(), MsgPacketStatus.RESPONSE_SUCCESS);
+        session.sendJsonMsg(keyMap, ActionType.GET_WEBSITE.name(), IdUtil.getInt(), MsgPacketStatus.SEND_REQUEST, msgPacket -> {
+            Map map = new Gson().fromJson(msgPacket.getDataStr(), Map.class);
+            Map<String, Object> response = new HashMap<>();
+            try {
+                PublicInfo responseSync = session.getResponseSync(ContentType.JSON, new HashMap<>(), ActionType.LOAD_PUBLIC_INFO, PublicInfo.class);
+                map.put("displayName", responseSync.getTitle());
+                MailUtil.sendMail(map.get("to").toString(), "这是一封测试邮件", "<div>当你看到这封邮件的时候，说明邮件服务已经可以正常工作了</h3>\n", map, new ArrayList<>());
+                response.put("status", 200);
+            } catch (Exception e) {
+                LOGGER.log(Level.SEVERE, "send email error ", e);
+                response.put("status", 500);
             }
+            session.sendMsg(ContentType.JSON, response, requestPacket.getMethodStr(), requestPacket.getMsgId(), MsgPacketStatus.RESPONSE_SUCCESS);
         });
-
     }
 }
